Updated the dashboard routing, and component
