--- prime-angular-frontend/src/app/core/models/token.model.ts
@@ -1,4 +1,0 @@
-export interface Token {
-  name: string;
-  at_hash: string;
-}
