--- prime-angular-frontend/src/app/dashboard/components/dashboard/dashboard.component.scss
@@ -1,70 +1,0 @@
-@import './palette';
-
-// Sidenav, toolbar, and icons included in the dashboard are styled using global
-// styles to avoid issues with the shadow DOM of components preventing styling.
-// @see src/_material-fixes.scss
-
-.brand {
-  width: 50px;
-}
-
-.center-brand {
-  margin: 0 auto;
-}
-
-.container {
-  // Provide space between the end of a view, and
-  // the bottom of the scrollable content
-  padding-bottom: 3.5rem;
-  padding-left: 1.65rem;
-}
-
-// .mat-nav-list .active {
-//   color: theme-palette(green);
-// }
-
-.st0 {
-  fill-rule:evenodd;
-  clip-rule:evenodd;
-  fill: theme-palette(blue);
-}
-
-.st1 {
-  fill:none;
-}
-
-.logo {
-  width: 4em;
-  height: 4em;
-
-  margin: 1.2em auto 1.5em;
-  display: block;
-}
-
-.mat-drawer {
-  background-color: #eeeeee !important;
-}
-
-.bg::after {
-  content: "";
-  background: url("../../../../assets/pharmacy-background.jpg");
-  opacity: 0.3;
-  top: 0;
-  left: 0;
-  bottom: 0;
-  right: 0;
-  position: absolute;
-  z-index: -1;
-}
-
-.text-blue {
-  color: theme-palette(blue, lighter) !important;
-}
-
-.text-black {
-  color: theme-palette(blue) !important;
-}
-
-.hide {
-  display: none;
-}
