+++ prime-angular-frontend/src/app/dashboard/shared/components/dashboard/dashboard.component.html
@@ -1,0 +1,1 @@
+<router-outlet></router-outlet>
