+++ prime-angular-frontend/src/app/dashboard/shared/components/dashboard/dashboard.component.ts
@@ -1,0 +1,12 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-dashboard',
+  templateUrl: './dashboard.component.html',
+  styleUrls: ['./dashboard.component.scss']
+})
+export class DashboardComponent implements OnInit {
+  constructor() { }
+
+  public ngOnInit() { }
+}
